Generate a static site for documentation with embedded profiles (#38)
The short of the patch is that there's now a website for the
documentation for the project. http://mpickering.github.io/eventlog2json

The files to generate the site are in the docs/ folder. Modify those and
the site is automatically deployed when the changes are merged to
master.

Example profiles can be embedded into the site by using a code block.

```{.eventlog}
path/to/eventlog
```

The long of the patch is that this required a bit of nix-wrangling to
achieve in a sensible way but shouldn't be too bad to maintain now its
set up.