Add options to filter traces
The -i option matches a substring against a trace and includes it on the
chart if the trace matches.

The -x option matches a substring against a trace and removes it on a
match.

Combining together the -i and -x option prefers the -i option. For
example, if a trace matches the -i option then it will be displayed.