Use UserMarker rather than UserMessage events.
UserMarker events are created using `traceMarker` which is intended for
low-frequency events and for visual display in tools such as
eventlog2html and threadscope. This means that you should now use
`traceMarker` in order for events to appear on the chart.

If your program uses `traceEvent` rather than `traceMarker` then you can
use the `--include-trace-events` flag to recover the old behaviour.

Fixes #94