Continue refactoring
The general idea is that each band can set both a short and long
description. The short description is displayed on hover and in the
legend. The long description is displayed underneath the chart in full.

The motivation is to be able to display the full cost centre stack.

At the moment there should be no change in functionality.