Fix start frame time
For some reason the events in the eventlog are not ordered in all
situations. Therefore taking the time of the first event doesn't give
you the actual start time. I don't see any reason why the start time
should be anything other than 0 so I just hard coded it in now.

Thanks to Andres for pointing out another bug.